Move the actual writing from ConnectionInner to RustConnection
This commit moves the actual writing of the request bytes "to the wire"
from ConnectionInner to RustConnection. ConnectionInner now only manages
the state and the actual I/O is done by RustConnection.
