+++ src/rust_connection/inner.rs
@@ -2,7 +2,7 @@
 
 use std::collections::VecDeque;
 use std::convert::TryInto;
-use std::io::{ErrorKind, IoSlice, Read, Write};
+use std::io::{Read, Write};
 
 use super::RawEventAndSeqNumber;
 use crate::connection::{DiscardMode, RequestKind, SequenceNumber};
@@ -33,7 +33,7 @@ where
 {
     // The underlying byte stream used for writing to the X11 server. Reading is done outside of
     // this struct (for synchronisation reasons).
-    write: W,
+    pub(crate) write: W,
 
     // The sequence number of the last request that was written
     last_sequence_written: SequenceNumber,
@@ -127,7 +127,6 @@ where
     /// should try again.
     pub(crate) fn send_request(
         &mut self,
-        bufs: &[IoSlice<'_>],
         kind: RequestKind,
     ) -> Result<Option<SequenceNumber>, std::io::Error> {
         if self.next_reply_expected + SequenceNumber::from(u16::max_value())
@@ -150,34 +149,6 @@ where
             discard_mode: None,
         };
         self.sent_requests.push_back(sent_request);
-
-        // Now actually send the buffers
-        // FIXME: We must always be able to read when we write
-        let mut bufs = bufs;
-        while !bufs.is_empty() {
-            let mut count = self.write.write_vectored(bufs)?;
-            if count == 0 {
-                return Err(std::io::Error::new(
-                    ErrorKind::WriteZero,
-                    "failed to write anything",
-                ));
-            }
-            while count > 0 {
-                if count >= bufs[0].len() {
-                    count -= bufs[0].len();
-                } else {
-                    let remaining = &bufs[0][count..];
-                    self.write.write_all(remaining)?;
-                    count = 0;
-                }
-                bufs = &bufs[1..];
-
-                // Skip empty slices
-                while bufs.first().map(|s| s.len()) == Some(0) {
-                    bufs = &bufs[1..];
-                }
-            }
-        }
 
         Ok(Some(seqno))
     }
@@ -396,8 +367,6 @@ fn read_setup(read: &mut impl Read) -> Result<Setup, ConnectError> {
 
 #[cfg(test)]
 mod test {
-    use std::io::IoSlice;
-
     use super::{read_setup, ConnectionInner};
     use crate::connection::RequestKind;
     use crate::errors::ConnectError;
@@ -470,39 +439,25 @@ mod test {
         // The connection must send a sync (GetInputFocus) request every 2^16 requests (that do not
         // have a reply). Thus, this test sends more than that and tests for the sync to appear.
 
-        let length = 1u16.to_ne_bytes();
-        let no_operation = [127, 0, length[0], length[1]];
-        let get_input_focus = [43, 0, length[0], length[1]];
-
         // Set up a connection that writes to this array
         let mut written = [0; 0x10000 * 4 + 4];
         let mut output = &mut written[..];
         let mut connection = ConnectionInner::new(&mut output);
 
         for num in 1..0x10000 {
-            let seqno =
-                connection.send_request(&[IoSlice::new(&no_operation)], RequestKind::IsVoid)?;
+            let seqno = connection.send_request(RequestKind::IsVoid)?;
             assert_eq!(Some(num), seqno);
         }
         // request 0x10000 should be a sync, hence the next one is 0x10001
-        let seqno = connection.send_request(&[IoSlice::new(&no_operation)], RequestKind::IsVoid)?;
+        let seqno = connection.send_request(RequestKind::IsVoid)?;
         assert_eq!(None, seqno);
 
-        let seqno = connection.send_request(&[IoSlice::new(&get_input_focus)], RequestKind::HasResponse)?;
+        let seqno = connection.send_request(RequestKind::HasResponse)?;
         assert_eq!(Some(0x10000), seqno);
 
-        let seqno = connection.send_request(&[IoSlice::new(&no_operation)], RequestKind::IsVoid)?;
+        let seqno = connection.send_request(RequestKind::IsVoid)?;
         assert_eq!(Some(0x10001), seqno);
 
-        let mut expected: Vec<_> = std::iter::repeat(&no_operation)
-            .take(0xffff)
-            .flatten()
-            .copied()
-            .collect();
-        expected.extend_from_slice(&get_input_focus);
-        expected.extend_from_slice(&no_operation);
-
-        assert_eq!(&written[..], &expected[..]);
         Ok(())
     }
 
@@ -511,26 +466,16 @@ mod test {
         // Compared to the previous test, this uses RequestKind::HasResponse, so no sync needs to
         // be inserted.
 
-        let length = 1u16.to_ne_bytes();
-        let get_input_focus = [43, 0, length[0], length[1]];
-
         // Set up a connection that writes to this array
         let mut written = [0; 0x10001 * 4];
         let mut output = &mut written[..];
         let mut connection = ConnectionInner::new(&mut output);
 
         for num in 1..=0x10001 {
-            let seqno = connection
-                .send_request(&[IoSlice::new(&get_input_focus)], RequestKind::HasResponse)?;
+            let seqno = connection.send_request(RequestKind::HasResponse)?;
             assert_eq!(Some(num), seqno);
         }
 
-        let expected: Vec<_> = std::iter::repeat(&get_input_focus)
-            .take(0x10001)
-            .flatten()
-            .copied()
-            .collect();
-        assert_eq!(&written[..], &expected[..]);
         Ok(())
     }
 
@@ -540,66 +485,19 @@ mod test {
         // the next request. Then it sends a RequestKind::HasResponse request so that no sync is
         // necessary. This is a regression test: Once upon a time, an unnecessary sync was done.
 
-        let length = 1u16.to_ne_bytes();
-        let no_operation = [127, 0, length[0], length[1]];
-        let get_input_focus = [43, 0, length[0], length[1]];
-
         // Set up a connection that writes to this array
         let mut written = [0; 0x10000 * 4];
         let mut output = &mut written[..];
         let mut connection = ConnectionInner::new(&mut output);
 
         for num in 1..0x10000 {
-            let seqno = connection
-                .send_request(&[IoSlice::new(&no_operation)], RequestKind::IsVoid)?;
+            let seqno = connection.send_request(RequestKind::IsVoid)?;
             assert_eq!(Some(num), seqno);
         }
 
-        let seqno = connection
-            .send_request(&[IoSlice::new(&get_input_focus)], RequestKind::HasResponse)?;
+        let seqno = connection.send_request(RequestKind::HasResponse)?;
         assert_eq!(Some(0x10000), seqno);
 
-        let mut expected: Vec<_> = std::iter::repeat(&no_operation)
-            .take(0xffff)
-            .flatten()
-            .copied()
-            .collect();
-        expected.extend_from_slice(&get_input_focus);
-
-        assert_eq!(&written[..], &expected[..]);
         Ok(())
-    }
-
-    fn partial_write_test(request: &[u8], expected_err: &str) {
-        let mut written = [0x21; 2];
-        let mut output = &mut written[..];
-        let mut connection = ConnectionInner::new(&mut output);
-        let request = [IoSlice::new(&request), IoSlice::new(&request)];
-        let error = connection
-            .send_request(&request, RequestKind::IsVoid)
-            .unwrap_err();
-        assert_eq!(expected_err, error.to_string());
-    }
-
-    #[test]
-    fn partial_write_larger_slice() {
-        partial_write_test(&[0; 4], "failed to write whole buffer");
-    }
-
-    #[test]
-    fn partial_write_slice_border() {
-        partial_write_test(&[0; 2], "failed to write anything");
-    }
-
-    #[test]
-    fn full_write_trailing_empty() {
-        let mut written = [0; 4];
-        let mut output = &mut written[..];
-        let mut connection = ConnectionInner::new(&mut output);
-        let (request1, request2) = ([0; 4], [0; 0]);
-        let request = [IoSlice::new(&request1), IoSlice::new(&request2)];
-        let _ = connection
-            .send_request(&request, RequestKind::IsVoid)
-            .unwrap();
     }
 }
